scripts/import/br/requirements.txt
python-slugify[unidecode]==4.0.1
requests==2.25.1
pandas==1.1.5
